+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetFormFactory/PaymentSheetFormFactoryConfig.swift
@@ -127,6 +127,15 @@ enum PaymentSheetFormFactoryConfig {
         }
     }
 
+    var opensCardScannerAutomatically: Bool {
+        switch self {
+        case .customerSheet(let customerSheetConfiguration):
+            return customerSheetConfiguration.opensCardScannerAutomatically
+        case .paymentElement(let paymentElementConfiguration):
+            return paymentElementConfiguration.opensCardScannerAutomatically
+        }
+    }
+
     func termsDisplayFor(paymentMethodType: PaymentSheet.PaymentMethodType) -> PaymentSheet.TermsDisplay {
         switch self {
         case .paymentElement(let configuration):