+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Views/CardScanningView.swift
@@ -26,15 +26,16 @@ private class CardScanningEasilyTappableButton: UIButton {
 
 /// For internal SDK use only
 @available(macCatalyst 14.0, *)
-@objc protocol STP_Internal_CardScanningViewDelegate: NSObjectProtocol {
-    func cardScanningView(
-        _ cardScanningView: CardScanningView, didFinishWith cardParams: STPPaymentMethodCardParams?)
+protocol STP_Internal_CardScanningViewDelegate: AnyObject {
+    func cardScanningView(_ cardScanningView: CardScanningView, didScanCard cardParams: STPPaymentMethodCardParams)
+    func cardScanningViewShouldClose(_ cardScanningView: CardScanningView)
 }
 
 /// For internal SDK use only
 @objc(STP_Internal_CardScanningView)
 @available(macCatalyst 14.0, *)
-class CardScanningView: UIView, STPCardScannerDelegate {
+class CardScanningView: UIView {
+
     private(set) weak var cameraView: STPCameraView?
 
     weak var delegate: STP_Internal_CardScanningViewDelegate?
@@ -51,25 +52,14 @@ class CardScanningView: UIView, STPCardScannerDelegate {
         }
     }
 
-    func cardScanner(
-        _ scanner: STPCardScanner,
-        didFinishWith cardParams: STPPaymentMethodCardParams?,
-        error: Error?
-    ) {
-        if error != nil {
-            self.isDisplayingError = true
-        } else {
-            self.delegate?.cardScanningView(self, didFinishWith: cardParams)
-        }
-    }
-
     private lazy var cardScanner: STPCardScanner? = nil
 
-    private static let cardSizeRatio: CGFloat = 2.125 / 3.370  // ID-1 card size (in inches)
+    private static let cardSizeRatio: CGFloat = 2.125 / 3.370  // ID-1 card size (in inches) converted to ratio
     private static let cardCornerRadius: CGFloat = 0.125 / 3.370  // radius / ID-1 card width
     private static let cornerRadius: CGFloat = 4
     private static let cardInset: CGFloat = 32
-    private static let textInset: CGFloat = 14
+    private static let errorLabelInset: CGFloat = 8
+    private static let closeButtonInset: CGFloat = 8
 
     private lazy var cardOutlineView: UIView = {
         let view = UIView()
@@ -121,12 +111,6 @@ class CardScanningView: UIView, STPCardScannerDelegate {
         return button
     }()
 
-    private lazy var containerView: UIView = {
-        let containerView = UIView()
-        containerView.translatesAutoresizingMaskIntoConstraints = false
-        return containerView
-    }()
-
     private func setupBlurView() {
         let vibrancyEffect = UIVibrancyEffect(blurEffect: blurEffect)
         let vibrancyEffectView = UIVisualEffectView(effect: vibrancyEffect)
@@ -154,19 +138,17 @@ class CardScanningView: UIView, STPCardScannerDelegate {
         ])
     }
 
-    func start() {
+    func startScanner() {
         cardScanner?.start()
     }
 
-    func stop() {
-        if isDisplayingError {
-            self.delegate?.cardScanningView(self, didFinishWith: nil)
-        }
-        cardScanner?.stop()
+    func stopScanner() {
+        cardScanner?.stop(didError: false)
     }
 
     @objc private func closeTapped() {
-        self.stop()
+        delegate?.cardScanningViewShouldClose(self)
+        stopScanner()
     }
 
     var snapshotView: UIView?
@@ -174,7 +156,7 @@ class CardScanningView: UIView, STPCardScannerDelegate {
     // The shape layers don't animate cleanly during setHidden,
     // so let's use a snapshot view instead.
     func prepDismissAnimation() {
-        if let snapshot = self.containerView.snapshotView(afterScreenUpdates: true) {
+        if let snapshot = snapshotView(afterScreenUpdates: true) {
             self.addSubview(snapshot)
             self.snapshotView = snapshot
         }
@@ -199,14 +181,13 @@ class CardScanningView: UIView, STPCardScannerDelegate {
 
         closeButton.addTarget(self, action: #selector(closeTapped), for: .touchUpInside)
 
-        self.addSubview(containerView)
-        containerView.addSubview(cameraView)
-        containerView.addSubview(cardOutlineView)
-        containerView.addSubview(cardOuterBlurView)
-        containerView.addSubview(errorLabel)
-        containerView.addSubview(closeButton)
+        self.addSubview(cameraView)
+        self.addSubview(cardOutlineView)
+        self.addSubview(cardOuterBlurView)
+        self.addSubview(errorLabel)
+        self.addSubview(closeButton)
 
-        containerView.layer.cornerRadius = CardScanningView.cornerRadius
+        self.layer.cornerRadius = Self.cornerRadius
         self.cameraView = cameraView
         cameraView.layer.cornerRadius = CardScanningView.cornerRadius
         self.cameraView?.translatesAutoresizingMaskIntoConstraints = false
@@ -216,46 +197,43 @@ class CardScanningView: UIView, STPCardScannerDelegate {
         // To get the right animation, we'll add a breakable bottom constraint
         // and enable clipsToBounds. Then, when hidden, the view will shrink while
         // the contents remain pinned to the top.
-        let bottomConstraint = containerView.bottomAnchor.constraint(equalTo: self.bottomAnchor)
-        bottomConstraint.priority = .defaultHigh
+        let bottomConstraints = [
+            cameraView.bottomAnchor.constraint(equalTo: self.bottomAnchor),
+            cardOuterBlurView.bottomAnchor.constraint(equalTo: self.bottomAnchor),
+            cardOutlineView.bottomAnchor.constraint(equalTo: self.bottomAnchor, constant: -Self.cardInset),
+        ]
+        bottomConstraints.forEach {
+            $0.priority = .defaultHigh
+        }
         self.clipsToBounds = true
+        self.addConstraints(bottomConstraints)
 
         self.addConstraints(
             [
-                containerView.topAnchor.constraint(equalTo: self.topAnchor),
-                containerView.leftAnchor.constraint(equalTo: self.leftAnchor),
-                containerView.rightAnchor.constraint(equalTo: self.rightAnchor),
-                bottomConstraint,
+                cameraView.leftAnchor.constraint(equalTo: self.leftAnchor),
+                cameraView.rightAnchor.constraint(equalTo: self.rightAnchor),
+                cameraView.topAnchor.constraint(equalTo: self.topAnchor),
 
-                cameraView.bottomAnchor.constraint(equalTo: containerView.bottomAnchor, constant: 0),
-                cameraView.leftAnchor.constraint(equalTo: containerView.leftAnchor, constant: 0),
-                cameraView.rightAnchor.constraint(equalTo: containerView.rightAnchor, constant: 0),
-                cameraView.topAnchor.constraint(equalTo: containerView.topAnchor, constant: 0),
+                cardOuterBlurView.leftAnchor.constraint(equalTo: self.leftAnchor),
+                cardOuterBlurView.rightAnchor.constraint(equalTo: self.rightAnchor),
+                cardOuterBlurView.topAnchor.constraint(equalTo: self.topAnchor),
 
-                cardOuterBlurView.topAnchor.constraint(equalTo: containerView.topAnchor, constant: 0),
-                cardOuterBlurView.leftAnchor.constraint(equalTo: containerView.leftAnchor, constant: 0),
-                cardOuterBlurView.rightAnchor.constraint(equalTo: containerView.rightAnchor, constant: 0),
-                cardOuterBlurView.bottomAnchor.constraint(equalTo: containerView.bottomAnchor, constant: 0),
-
-                errorLabel.centerYAnchor.constraint(equalTo: containerView.centerYAnchor, constant: 0),
-                errorLabel.leftAnchor.constraint(equalTo: cardOutlineView.leftAnchor, constant: 8),
+                errorLabel.leftAnchor.constraint(equalTo: cardOutlineView.leftAnchor, constant: Self.errorLabelInset),
                 errorLabel.rightAnchor.constraint(
-                    equalTo: cardOutlineView.rightAnchor, constant: -8),
+                    equalTo: cardOutlineView.rightAnchor, constant: -Self.errorLabelInset),
+                errorLabel.centerYAnchor.constraint(equalTo: cardOutlineView.centerYAnchor),
 
-                closeButton.topAnchor.constraint(equalTo: containerView.topAnchor, constant: 8),
-                closeButton.rightAnchor.constraint(equalTo: containerView.rightAnchor, constant: -8),
+                closeButton.rightAnchor.constraint(equalTo: self.rightAnchor, constant: -Self.closeButtonInset),
+                closeButton.topAnchor.constraint(equalTo: self.topAnchor, constant: Self.closeButtonInset),
 
                 cardOutlineView.heightAnchor.constraint(
                     equalTo: cardOutlineView.widthAnchor, multiplier: CardScanningView.cardSizeRatio),
-
-                cardOutlineView.topAnchor.constraint(
-                    equalTo: containerView.topAnchor, constant: CardScanningView.cardInset),
                 cardOutlineView.leftAnchor.constraint(
-                    equalTo: containerView.leftAnchor, constant: CardScanningView.cardInset),
+                    equalTo: self.leftAnchor, constant: CardScanningView.cardInset),
                 cardOutlineView.rightAnchor.constraint(
-                    equalTo: containerView.rightAnchor, constant: -CardScanningView.cardInset),
-                cardOutlineView.bottomAnchor.constraint(
-                    equalTo: containerView.bottomAnchor, constant: -CardScanningView.cardInset),
+                    equalTo: self.rightAnchor, constant: -CardScanningView.cardInset),
+                cardOutlineView.topAnchor.constraint(
+                    equalTo: self.topAnchor, constant: CardScanningView.cardInset),
             ])
     }
 
@@ -288,9 +266,25 @@ class CardScanningView: UIView, STPCardScannerDelegate {
 
     override func willMove(toWindow newWindow: UIWindow?) {
         if newWindow == nil {
-            stop()
+            stopScanner()
         }
         super.willMove(toWindow: newWindow)
     }
 }
+
+@available(macCatalyst 14.0, *)
+extension CardScanningView: STPCardScannerDelegate {
+
+    func cardScanner(_ scanner: STPCardScanner, didFinishWith cardParams: STPPaymentMethodCardParams?) {
+        if let cardParams {
+            self.delegate?.cardScanningView(self, didScanCard: cardParams)
+        }
+        self.delegate?.cardScanningViewShouldClose(self)
+    }
+
+    func cardScannerDidError(_ scanner: STPCardScanner) {
+        self.isDisplayingError = true
+    }
+}
+
 #endif